Enable KnownBool solver
In a future release ghc-typelits-knownnat will introduce a KnownBool
analogue to KnownNat, allowing the solver to solve for applications of
If. This allows the Min and Max type families in ghc-typelits-extra to
be expressed as a single equation, making them more amenable to solving
via SMT solver.

However, after this change solving for Min and Max requires the
KnownBool solver.

See ghc-typelits-knownnat#23.
See ghc-typelits-extra#16.